requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
Django==1.10.8
django-braces==1.11.0
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
enum34==1.1.6
flake8==3.4.1
idna==2.6
mccabe==0.6.1
MySQL-python==1.2.5
oauthlib==2.0.4
pycodestyle==2.3.1
pyflakes==1.5.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
